backend/requirements.txt
flask>=3.1.0
flask-cors>=6.0.0
flask-limiter>=4.1.0
PyJWT>=2.7.0
bcrypt>=3.2.0
